Added end points to journal entries as well

+++ src/main/java/com/edigest/journalApp/controller/JournalEntryController.java
@@ -8,11 +8,14 @@ import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @RestController
 @Component
@@ -25,9 +28,11 @@ public class JournalEntryController {
     private UserService userService;
 
 
-    @GetMapping("/{userName}")
-    public ResponseEntity<?> getAllJournalEntriesOfUser(@PathVariable String userName) {
+    @GetMapping
+    public ResponseEntity<?> getAllJournalEntriesOfUser() {
         try {
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            String userName = authentication.getName();
             User user = userService.findUserByUserName(userName);
             if (user == null) {
                 return new ResponseEntity<>("User not found", HttpStatus.NOT_FOUND);
@@ -43,9 +48,11 @@ public class JournalEntryController {
     }
 
 
-    @PostMapping("{userName}")
-    public ResponseEntity<?> createJournalEntry(@RequestBody JournalEntry entry, @PathVariable String userName) {
+    @PostMapping()
+    public ResponseEntity<?> createJournalEntry(@RequestBody JournalEntry entry) {
         try {
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            String userName = authentication.getName();
             journalEntryService.saveJournalEntry(entry, userName);
             return new ResponseEntity<>(entry, HttpStatus.CREATED);
         } catch (Exception exception) {
@@ -56,22 +63,36 @@ public class JournalEntryController {
     @GetMapping("/id/{myId}")
     public ResponseEntity<JournalEntry> findById(@PathVariable ObjectId myId) {
         try {
-            Optional<JournalEntry> journalEntry = journalEntryService.findById(myId);
-            if (journalEntry.isPresent()) {
-                return new ResponseEntity<>(journalEntry.get(), HttpStatus.OK);
-            } else {
-                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            String userName = authentication.getName();
+            User user = userService.findUserByUserName(userName);
+            List<JournalEntry> collect = user.getJournalEntryList().stream().filter(x -> x.getId().equals(myId)).collect(Collectors.toList());
+            if(!collect.isEmpty())
+            {
+                Optional<JournalEntry> journalEntry = journalEntryService.findById(myId);
+                if (journalEntry.isPresent()) {
+                    return new ResponseEntity<>(journalEntry.get(), HttpStatus.OK);
+                }
             }
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-    @PutMapping("{userName}/{id}")
-    public ResponseEntity<?> updateJournalById(@RequestBody JournalEntry newEntry, @PathVariable String userName,
-                                               @PathVariable ObjectId id) {
+    @PutMapping("/{myId}")
+    public ResponseEntity<?> updateJournalById(@RequestBody JournalEntry newEntry, @PathVariable ObjectId myId) {
         try {
-            Optional<JournalEntry> journalEntry = journalEntryService.findById(id);
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+            String userName = authentication.getName();
+            User user = userService.findUserByUserName(userName);
+            List<JournalEntry> collect = user.getJournalEntryList().stream().filter(x -> x.getId().equals(myId)).collect(Collectors.toList());
+            if(collect.isEmpty())
+            {
+                return new ResponseEntity<>("Content not found",HttpStatus.NO_CONTENT);
+            }
+            Optional<JournalEntry> journalEntry = journalEntryService.findById(myId);
             JournalEntry oldEntry = journalEntry.get();
             if (journalEntry.isPresent()) {
                 oldEntry.setTitle(newEntry.getTitle() != null && !newEntry.getTitle().equals("") ?
@@ -86,8 +107,16 @@ public class JournalEntryController {
         }
     }
 
-    @DeleteMapping("{userName}/{myId}")
-    public ResponseEntity<?> deleteJournalById(@PathVariable ObjectId myId, @PathVariable String userName) {
+    @DeleteMapping("/id/{myId}")
+    public ResponseEntity<?> deleteJournalById(@PathVariable ObjectId myId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userName = authentication.getName();
+        User user = userService.findUserByUserName(userName);
+        List<JournalEntry> collect = user.getJournalEntryList().stream().filter(x -> x.getId().equals(myId)).collect(Collectors.toList());
+        if(collect.isEmpty())
+        {
+            return new ResponseEntity<>("No entries found to be deleted",HttpStatus.NO_CONTENT);
+        }
         Optional<JournalEntry> entryToDelete = journalEntryService.findById(myId);
         if (entryToDelete.isPresent()) {
             journalEntryService.deleteJournalById(myId, userName);
